Wrap application in context container
Originally the `renderRoot` property in LuxConfig was going to be a CSS
selector but it was changed to a DOM element to make it easier for lux
and more flexible for users.

ResponseModel was changed to bring fetch response properties along with
response data. Response data was also moved to body.

Make the default path string '/'

The `actions` array should always (probably) be included in API resources
but it shouldn't break the application if they aren't there.

The `Fatal` error page is a last hope to display an error by removing
as much of the framework from the equasion as possible. If an error is
thrown in the `Fatal` component there is no hope for rendering the app.

Start using the ReactJS feature "context" to automatically pass-down
a context object to all components within the application to enable
them to optionally use a state update function `setState`. This will
replace the need for an external state management library such as
Redux, MobX, etc.